Add the manifest, background and options files
In this commit, we are adding a bunch of files that implement the
basic functionality of the extension.

There might be some bugs to be fixed, and code to be improved.

When the user first installs the extension, the background script is
going to add an "!" on the extension icon, inviting the user to
click on it.

By clicking on the extension icon, the user will be redirected to the
extension's options page.

In this page, the user can provide an access token for their GitHub
account.

After providing the access token, the background script is going to
fetch the the count of PRs waiting for review on GitHub and display
this number on the extension icon.

Next time the user clicks on the extension icon, they will be
redirected to GitHub.