Fix error handling and add an offline behavior
In this commit, we are fixing the error handling so the credentials are
erased only if the API returns 401.

We are also changing the status of the badge when there is no internet
connection.

And finally, we are fixing the multiple listeners to the alarm by
ensuring we remove the listener before adding a new one.